backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
openpyxl==3.1.5
pdfplumber==0.11.4
python-jose[cryptography]==3.3.0
